Split info_panel.rs monolith into composable sub-modules
Break the 2609-LOC info_panel.rs into 8 focused sub-modules while
preserving the exact same public API:

- mod.rs: main info_panel_ui system, shared types (MinimapCache,
  InfoPanelExtras), and helper functions (format_pop, demand_bar, etc.)
- building_inspection.rs: building_inspection_ui and citizen display helpers
- policies.rs: policies_ui
- event_journal.rs: event_journal_ui and event_type_color
- advisor.rs: advisor_window_ui
- budget.rs: budget_panel_ui and budget_line helper
- keybinds.rs: panel_keybinds, quick_save_load_keybinds
- minimap.rs: build_minimap_pixels and minimap constants
- groundwater_tooltip.rs: groundwater_tooltip_ui

Closes #1162
